# api/openapi.yaml
openapi: 3.0.3
info:
  title: Go.Kitchen API (Go.Kitchen Service)
  description: |
    REST API сервиса доставки еды «Go.Kitchen».
    Предоставляет интерфейсы для покупателей (каталог заведений, меню, оформление заказов, отслеживание и отмена),
    а также для ресторанов-партнеров (управление меню, стоп-лист, остатки, очередь заказов, изменение статусов готовки и вебхуки).
  version: 1.0.0
  contact:
    name: Go.Kitchen Core Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Основной сервис Go.Kitchen (Core Kitchen Service)
  - url: http://localhost:8081
    description: Демо-сервис ресторана-партнера (Partner Restaurant Simulator)
...
paths:
...
  /api/v1/partner/orders/{order_id}/status:
    patch:
      operationId: updateOrderStatus
      security:
        - PartnerApiKey: []
      tags:
        - Partner API
      summary: Изменить статус заказа (жизненный цикл на кухне)
      description: |
        Переводит заказ по стейт-машине:
        `accepted` (принят) -> `cooking` (готовится) -> `ready` (готов) -> `delivering` (передан курьеру) -> `completed` (выдан).
        Если статус меняется на `cancelled`, остатки автоматически возвращаются на склад.
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '200':
          description: Статус заказа успешно изменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          description: Недопустимый переход статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhook:
    servers:
      - url: http://localhost:8081
        description: Демо-сервис ресторана-партнера (Partner Restaurant Simulator)
    post:
      operationId: receiveWebhook
      tags:
        - Partner Simulator
      summary: Прием входящих вебхуков от Авито.Кухни (симулятор партнера)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Вебхук принят и обработан
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "acknowledged"
        '400':
          description: Тело запроса не разобрано как JSON
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "invalid payload"
...
components:
  securitySchemes:
    AdminBootstrapToken:
      type: apiKey
      in: header
      name: X-Bootstrap-Token
      description: |
        Токен площадки, разрешающий поставить новое заведение на учёт. Отдельный от партнёрского
        ключа: тот отвечает на вопрос «какое я заведение», этот — «можно ли добавить ещё одно».
        В стенде `docker compose` задан как `demo-bootstrap-token-000000000001`.
    PartnerApiKey:
      type: http
      scheme: bearer
      description: |
        Закрытый доступ для заведений. Ключ выдаётся один раз при регистрации
        (`POST /api/v1/partner/restaurants`) и определяет, от имени какого заведения
        выполняется запрос. Чужие категории, блюда и заказы отвечают `404`, чтобы по API
        нельзя было перебирать идентификаторы конкурентов.
        Альтернативно ключ принимается в заголовке `X-Api-Key`.
  schemas:
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        menu_item_id:
          type: string
          format: uuid
        item_name:
          type: string
          example: "Чизбургер"
        price_cents:
          type: integer
          format: int64
          example: 25000
        quantity:
          type: integer
          example: 2
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, accepted, cooking, ready, delivering, completed, cancelled]
          example: "accepted"
        total_price_cents:
          type: integer
          format: int64
          example: 50000
        delivery_address:
          type: string
          example: "г. Москва, ул. Арбат, д. 12, кв. 44"
        comment:
          type: string
          example: "Пожалуйста, положите салфетки"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status_history:
          type: array
          description: Аудит переходов статуса от старых к новым.
          items:
            $ref: '#/components/schemas/OrderStatusHistoryEntry'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderStatusHistoryEntry:
      type: object
      description: "Запись аудита: кто и когда перевёл заказ в новый статус."
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        from_status:
          type: string
          description: Пустая строка для самой первой записи (создание заказа).
          example: "created"
        to_status:
          type: string
          enum: [created, accepted, cooking, ready, delivering, completed, cancelled]
          example: "accepted"
        actor:
          type: string
          enum: [customer, restaurant, system]
          description: Инициатор перехода.
          example: "restaurant"
        comment:
          type: string
          example: "Заказ подтвержден рестораном и передан шеф-повару"
        created_at:
          type: string
          format: date-time
...
    UpdateOrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [accepted, cooking, ready, delivering, completed, cancelled]
          example: "cooking"
        comment:
          type: string
          maxLength: 1000
          example: "Повар приступил к сборке бургеров"
    WebhookPayload:
      type: object
      properties:
        event:
          type: string
          example: "order.created"
        timestamp:
          type: string
          format: date-time
        data:
          type: object
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "requested resource not found"
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
  responses:
    UnauthorizedError:
      description: Отсутствует или неверен партнёрский API-ключ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "missing or invalid credentials"
    BadRequestError:
      description: Некорректный запрос или невалидный синтаксис данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "BAD_REQUEST"
              message: "invalid request data or syntax"
...
    NotFoundError:
      description: Запрашиваемый ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "requested resource not found"